backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0
python-dotenv>=1.0.0
httpx>=0.26.0
httpcore>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
aiofiles>=23.0.0
email-validator>=2.0.0
